+++ dnapayments/src/main/java/com/dna/sdk/dnapayments/api/service/AuthService.kt
@@ -5,6 +5,7 @@ import retrofit2.Response
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
+import java.math.BigDecimal
 
 internal interface AuthService {
     @FormUrlEncoded
@@ -15,7 +16,7 @@ internal interface AuthService {
         @Field("client_id") clientId: String,
         @Field("client_secret") clientSecret: String,
         @Field("invoiceId") invoiceId: String,
-        @Field("amount") amount: Double,
+        @Field("amount") amount: BigDecimal,
         @Field("currency") currency: String,
         @Field("terminal") terminal: String,
         @Field("paymentFormURL") paymentFormURL: String,
